*Fefactor subshader / master node generation to be more uniform *Add subshader generator as a 'real' thing, need to hook up UI now so that these can be added and removed.
