--- MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/HDPipeline/IUnlitSubShader.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace UnityEditor.ShaderGraph {
-    public interface IUnlitSubShader
-    {
-        string GetSubshader(UnlitMasterNode masterNode, GenerationMode mode);
-    }
-}
