+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/AlphaMode.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace UnityEditor.ShaderGraph {
+    public enum AlphaMode
+    {
+        Opaque,
+        AlphaBlend,
+        AdditiveBlend
+    }
+}
