+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/AlphaMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 05d113c37ee6ad24cbd4b11bbaab371f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
