+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/IPBRSubShader.cs
@@ -1,0 +1,7 @@
+using System;
+
+namespace UnityEditor.ShaderGraph {
+    public interface IPBRSubShader : ISubShader
+    {
+    }
+}
