+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/IPBRSubShader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8ba3702335c91a140bc0ee4c0891b160
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
