+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/ISubShader.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace UnityEditor.ShaderGraph {
+    public interface ISubShader
+    {
+        string GetSubshader(IMasterNode masterNode, GenerationMode mode);
+    }
+}
