+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/ISubShader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9e131b3f1bf111d4caab49dcc291757b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
