+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/MasterNodes/IUnlitSubShader.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace UnityEditor.ShaderGraph
+{
+    public interface IUnlitSubShader : ISubShader
+    {}
+}
