+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/Nodes/AbstractMaterialNode.cs
@@ -6,15 +6,6 @@ using UnityEditor.Graphing;
 
 namespace UnityEditor.ShaderGraph
 {
-    public static class GuidEncoder
-    {
-        public static string Encode(Guid guid)
-        {
-            string enc = Convert.ToBase64String(guid.ToByteArray());
-            return String.Format("{0:X}", enc.GetHashCode());
-        }
-    }
-
     [Serializable]
     public abstract class AbstractMaterialNode : INode, ISerializationCallbackReceiver, IGenerateProperties
     {
