+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/Nodes/GuidEncoder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3cf71f51f0308524080bb4aba2c9bb81
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
