+++ MaterialGraphProject/Assets/UnityShaderEditor/Editor/Data/Nodes/IMasterNode.cs
@@ -4,7 +4,6 @@ namespace UnityEditor.ShaderGraph
 {
     public interface IMasterNode
     {
-        SurfaceMaterialOptions options { get; }
         string GetShader(GenerationMode mode, string name, out List<PropertyCollector.TextureInfo> configuredTextures);
     }
 }
